Extract Composio bridge support
